Handle common text references.
The spec seems to permit for packages to request data from the `element` where they're declared. This isn't a full implementation, but it points the way.

(A lot of stuff isn't the full implementation. `eagle.dtd` sets out some pretty clear rules on what's expected, optional, and permitted in various contexts, although a lot of it simply isn't used in the live files I'm handling.