Simplify class structure. Add `params`.
I'd originally envisaged the system as a factory, able to read XML and spit out THREEjs objects. It's become increasingly clear that we'll need a lot more metadata, so I adapted the system to return a `Board` object... and eventually the factory was just a wrapper around `new Board`. So I simplified it down to one class/namespace.