Add `getElementsByTagNameAndLayer`.
Initial work towards obtaining correct wire data for board outlines.

This will have to be heavily revised. Currently, it seems to get any wires, whether they're part of the board, or just a library definition. We'll need to implement a proper document parser that assigns elements to be drawn to layers, so we know which definitions to actually use, and what their true coordinates are.