Implement layer populating.
The renderer now assembles a set of layers from design rule data, and populates them according to XML data. Currently, data is read from <plain> and <signals>. The <elements> and <element> tags are not yet implemented, but when they are, that will populate the entire layer edifice with applicable data.