Polish polygon rendering.
For some reason, some polygons don't render. They're definitely exactly the same as polygons that render on other layers, and other polygons on the same layer render fine, but they do not appear. Even if nothing else gets rendered, the polygons don't appear, so it probably isn't a context pollution.