Fix issue with out-of-order boundaries.
As seen in `stick.brd`, if the board designer didn't create the board boundary wire definitions in the correct order, they wouldn't render correctly. In fact, you could define boundary wires to start from the same location and diverge! This would never be caught by the basic path sort that I implemented.

I've now set it up to actually convert the coordinates to run in a single direction while sorting the boundary wires, both in `_buildDepthEdges()` and in `renderBounds()`. It shouldn't be entirely necessary in `_buildDepthEdges()`, which is running off the same wire definitions that are modified in `renderBounds()`, but it can't hurt to use the most robust reordering mechanism available.