Implement more robust bounds generation.
This will automatically convert `circle`, `rectangle` and `polygon` elements into `wire` elements. This allows the code to recognise non-wire shapes, without requiring a lot of re-engineering. Wires are good, because they convert easily to line segments in a path - useful both for rendering to the canvas and for generating extruded edge geometry.