Improve bounds generation.
Implements a more robust algorithm for connecting bounds segments. It repeatedly looks for wires that will fit the current end of path, flipping them if necessary. If it can't find a candidate, it will start a new subpath by grabbing the next wire.

The previous algorithm turned out to be very fragile upon encountering direction reversals.