Improve bounds generation.
A flaw existed. If there was a discontinuity in bounds definitions, and the wires on the far side were declared in reverse, they would be unable to find a suitable connection in the loop.

The fix: also check for possible connections to the _front_ of the wire array. This allows a connectable wire segment to connect first time every time.

This has enabled at least one dodgy BRD to look perfect.