Fix polygon rendering.
`drawPolygons()` was not correctly respecting the handedness of curves. This could lead to extremely large filled arcs covering much or all of a board, curving corners being "cut out", and other unpleasant texture artifacts.

The main edit here is in `ctx.arc`, which now features `chordData.curve < 0` as its terminal parameter. This tells the canvas context whether the arc is clockwise or anticlockwise.