Improve font flexibility
As it turns out, my information was somewhat incorrect. The OCR-A font is not what EAGLE uses; that appears to be a proprietary font known only as "vector". Accordingly, I've updated the code to not require any sort of font by default, and made OCR-A an optional inclusion. The board will render using a monospace system font by default.