Update to THREE r79; implement compatibility fixes
THREE r79 introduces some refactoring to the `Path` class. This broke some arcs in board edge definitions. I don't think this is a bug in THREE; it looks like things are behaving more logically now.

This means, of course, that we were implementing bug-compensatory code. I disapprove of bug-compensatory code.

The fix removes a chunk of code that figures out the winding angle of the path and chooses whether to arc clockwise or counterclockwise. It looks like the `Path` can cope with negative angles without assistance now.

All sample board corners and arcs appear unbroken and correct.