Add workaround for Connector's inability to handle PCB rotation.
The Connector/connectTo code appears not to handle rotation of the PCB/parent object correctly.

It seems to not take into account the parent in its calculations. I've been able to modify it to set the position correctly but not the rotation. So, for now, using this work around which doesn't appear to have noticable side effects with the current example.

The symptom of the rotation not being handled is that the component models are placed in seemingly random (but actually dependent on the rotation of the board at the time the particular component model is loaded) "exploding" pattern of position & rotation.