Avoid attempting to populate footprints labeled "DNP" ("Do Not Populate).
Notes:

1) An element value of "DNP" is used for footprints on SparkFun boards to indicate
there is no component placed (e.g. when a component is intended as a custom
option for purchasers). It's not clear how wide-spread this convention is.

2) The current implementation means that the model file will be requested
even if all instances of a footprint are marked DNP. This approach was
chosen as the simplest change that would work but the additional optimisation
could be added if desired.