Rework SparkFun Redstick example to use map & models from standard library.
In order to keep a completely self-contained example I've copied a subset
of models and the mapping file from our component model standard library.

The files/folders have been named with a "-redstick" suffix to make clear
they are a subset of the entire library and enable the full library to be
symlinked into the examples folder if desired.

This also serves as an example of how to supply an alternate map file URL.