Fix orientation of Eagle-Loader Connectors on back side of boards.
Without this fix component models end up rotated 180 degrees the
wrong way on the backside of the board.

Before I realised this was an issue with the connectors rather than
the models I'd committed model orientation changes to the standard
library, so this fix requires use of an updated standard library.

The erroneous orientation fixes in the standard library resulted in
components on the top side being rotated incorrectly.