Fix handling of footprints with non-centre origins. (#24)
Previous code assumed that all component footprints were centered
around their origin but that turned out not to be the case. This
resulted in model files being placed incorrectly.

Now the component model map can specify `"center": false` when
a footprint and model have a non-centered origin.

The updates to the standard library to use this functionality
still need to be committed.

For more detail see: Issue #24.