Fix bug that meant URL prefix functionality didn't actually work.
While the code checked `activeComponentMapCfg.meta.urlPrefix` was present
the URL was built with `activeComponentMapCfg.urlPrefix` (note the
lack of `.meta` in the middle) which lead to the URL being
prefixed with "undefined" which isn't very uesful...

I think part of the reason this slipped through was because of
local caching when testing and need to choose a suitable prefix.
Since it seems GitHub will work for .STL file hosting in some
form I'll probably add an example demonstrating that--which will
also serve as a test for the URL prefix feature.