Fix bug where (invalid) Connector names were created for mechanical holes.
The intention was that Connector names should only be generated for pins/pads
but it turned out we were also generating names (which were invalid,
e.g. "S1.null") for mechanical holes. This fix ensures names are
only generated for pins/pads connectors.

(This fix should really be made on the parent of this branch.)